Narrow struct sctstr member sct_mobil to char
To bring it in line with unit mobility.

+++ src/lib/commands/setsect.c
@@ -235,7 +235,7 @@ setsector(void)
 		    current = 127;
 		pr("Mobility in %s changed to %d.\n",
 		   xyas(sect.sct_x, sect.sct_y, player->cnum), current);
-		sect.sct_mobil = (short)current;
+		sect.sct_mobil = current;
 		putsect(&sect);
 	    }
 	    break;
